+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # 2.x branch
 ## 2.26 branch
+### 2.26.3
+* added the Disqus comment system for pages and posts. By default, comments are
+    enabled for posts;
+* added `enable_comments` field to `Pages` and `Posts` tables.
+    Added `VersionUpdatesCommand::addEnableCommentsField()` method;
+* fixed bug for static pages with locales.
+
 ### 2.26.2
 * it is now possible to disable the sitemap for some content, using the MeCms
     configuration file;